service impl

+++ src/main/java/si/feri/mag/soa/smarthome/services/IAirConditionerService.java
@@ -13,7 +13,7 @@ public interface IAirConditionerService {
     @WebMethod
     boolean isAcOn();
     @WebMethod
-    int getAmbientTemperatureFromAc();
+    float getAmbientTemperatureFromAc();
     @WebMethod
     void setAcTimer(int timeMins);
     @WebMethod
